support for new testing path for special models
