+++ .github/workflows/test_iree.yml
@@ -101,23 +101,15 @@ jobs:
           - name: cpu_llvm_task
             runs-on: nodai-amdgpu-w7900-x86-64
             models-config-file: config_pytorch_models_cpu_llvm_task.json
-            sdxl-prompt-encoder-config-file: config_sdxl_prompt_encoder_cpu_llvm_task.json
-            sdxl-unet-config-file: config_sdxl_scheduled_unet_cpu_llvm_task.json
-            sdxl-vae-decode-config-file: config_sdxl_vae_decode_cpu_llvm_task.json
           - name: gpu_mi250_rocm
             runs-on: nodai-amdgpu-mi250-x86-64
             models-config-file: config_gpu_rocm_models.json
-            sdxl-prompt-encoder-config-file: config_sdxl_prompt_encoder_gpu_rocm.json
-            sdxl-unet-config-file: config_sdxl_scheduled_unet_gpu_rocm.json
-            sdxl-vae-decode-config-file: config_sdxl_vae_decode_gpu_rocm.json
     env:
       VENV_DIR: ${{ github.workspace }}/.venv
+      ROCM_CHIP: gfx90a
       IREE_TEST_FILES: ~/iree_tests_cache
       IREE_TEST_PATH_EXTENSION: ${{ github.workspace }}/iree_tests/specs
       MODELS_CONFIG_FILE_PATH: iree_tests/configs/${{ matrix.models-config-file }}
-      SDXL_PROMPT_ENCODER_CONFIG_FILE_PATH: iree_tests/configs/${{ matrix.sdxl-prompt-encoder-config-file }}
-      SDXL_UNET_CONFIG_FILE_PATH: iree_tests/configs/${{ matrix.sdxl-unet-config-file }}
-      SDXL_VAE_DECODE_CONFIG_FILE_PATH: iree_tests/configs/${{ matrix.sdxl-vae-decode-config-file }}
     steps:
       - name: "Checking out repository"
         uses: actions/checkout@v4
@@ -141,12 +133,19 @@ jobs:
         run: |
           source ${VENV_DIR}/bin/activate
           python3 -m pip install -r iree_tests/requirements.txt
+          pip install --no-compile --pre --upgrade -e iree_special_models
 
       # Download remote files.
       - name: "Downloading remote files for real weight model tests"
         run: |
           source ${VENV_DIR}/bin/activate
-          python3 iree_tests/download_remote_files.py --root-dir pytorch/models
+          python3 iree_tests/download_remote_files.py --root-dir iree_tests/pytorch/models
+      
+      # Download remote files.
+      - name: "Downloading remote files for special model tests"
+        run: |
+          source ${VENV_DIR}/bin/activate
+          python3 iree_tests/download_remote_files.py --root-dir iree_special_models
 
       - name: "Running real weights model tests"
         if: ${{ !cancelled() }}
@@ -162,57 +161,39 @@ jobs:
             --timeout=1200 \
             --durations=0 \
             --config-files=${MODELS_CONFIG_FILE_PATH}
-
-      - name: "Running real weights SDXL prompt encoder tests"
-        id: prompt_encoder
-        if: ${{ !cancelled() }}
+      
+      - name: "Running SDXL special model tests"
+        id: special_models_sdxl
+        if: |
+          ${{ !cancelled() }} &&
+          contains(matrix.name, 'rocm')
         run: |
           source ${VENV_DIR}/bin/activate
-          pytest iree_tests/pytorch/models/sdxl-prompt-encoder-tank \
+          pytest iree_special_models/sdxl \
             -rpfE \
-            -k real_weights \
-            --no-skip-tests-missing-files \
             --capture=no \
             --log-cli-level=info \
             --timeout=1200 \
-            --durations=0 \
-            --config-files=${SDXL_PROMPT_ENCODER_CONFIG_FILE_PATH}
-
-      - name: "Running real weights SDXL scheduled unet tests"
-        id: unet
-        if: ${{ !cancelled() }}
+            --durations=0
+      
+      - name: "Running SD3 special model tests"
+        id: special_models_sd3
+        if: |
+          ${{ !cancelled() }} &&
+          contains(matrix.name, 'rocm')
         run: |
           source ${VENV_DIR}/bin/activate
-          pytest iree_tests/pytorch/models/sdxl-scheduled-unet-3-tank \
+          pytest iree_special_models/sd3 \
             -rpfE \
-            -k real_weights \
-            --no-skip-tests-missing-files \
             --capture=no \
             --log-cli-level=info \
             --timeout=1200 \
-            --durations=0 \
-            --config-files=${SDXL_UNET_CONFIG_FILE_PATH}
-
-      - name: "Running real weights SDXL vae decode tests"
-        id: vae
-        if: ${{ !cancelled() }}
-        run: |
-          source ${VENV_DIR}/bin/activate
-          pytest iree_tests/pytorch/models/sdxl-vae-decode-tank \
-            -rpfE \
-            -k real_weights \
-            --no-skip-tests-missing-files \
-            --capture=no \
-            --log-cli-level=info \
-            --timeout=1200 \
-            --durations=0 \
-            --config-files=${SDXL_VAE_DECODE_CONFIG_FILE_PATH}
+            --durations=0
 
       - name: "Running SDXL ROCm pipeline benchmark"
         if: |
           contains(matrix.name, 'rocm') &&
-          steps.prompt_encoder.outcome == 'success' &&
-          steps.unet.outcome == 'success' && steps.vae.outcome == 'success'
+          steps.special_models_sdxl.outcome == 'success'
         run: |
           source ${VENV_DIR}/bin/activate
           pytest iree_tests/benchmarks/sdxl/benchmark_sdxl_rocm.py \
