+++ iree_special_models/ireers/__init__.py
@@ -1,0 +1,7 @@
+from .fixtures import (
+    IreeCompileException,
+    IreeRunException,
+    get_compile_cmd,
+    iree_compile,
+    iree_run_module,
+)
