+++ iree_special_models/sd3/vae-decode/test_cases.json
@@ -1,0 +1,13 @@
+{
+    "file_format": "test_cases_v0",
+    "test_cases": [
+      {
+        "name": "real_weights",
+        "remote_files": [
+          "https://sharkpublic.blob.core.windows.net/sharkpublic/sai/sd3-vae/inference_input.0.bin",
+          "https://sharkpublic.blob.core.windows.net/sharkpublic/sai/sd3-vae/inference_output.0.bin",
+          "https://sharkpublic.blob.core.windows.net/sharkpublic/sai/sd3-vae/real_weights.irpa"
+        ]
+      }
+    ]
+}
