--- iree_tests/pytorch/complex_models/ireers/__init__.py
@@ -1,5 +1,0 @@
-from .fixtures import (
-    get_compile_cmd,
-    iree_compile,
-    iree_run_module,
-)
