--- iree_tests/pytorch/complex_models/sd3/mmdit/model.mlirbc
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:83ba72246a5d83ee2a9849a70bf62295c8a79f88f16dee71eeb08bd0ba8cfe15
-size 430357
