--- iree_tests/pytorch/complex_models/sd3/prompt-encoder/model.mlirbc
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7cc1456e1b5e2e5138db5bc00421b9f8e46710745b5f847579c1036b1df56f7c
-size 969269
