--- iree_tests/pytorch/complex_models/sd3/vae-decode/model.mlirbc
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7ab1d958d33b3cc291623fe9e0244690d8c07d9b883dfe50401028274c0c0cc8
-size 69621
