--- iree_tests/pytorch/complex_models/sdxl/prompt-encoder/model.mlirbc
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:70b75a88e0f26588df0df4568b7498332e5baccdd862de9970ade2619610ae74
-size 693392
