--- iree_tests/pytorch/models/sdxl-prompt-encoder-tank/splat_data_flags.txt
@@ -1,5 +1,0 @@
---input="1x64xi64"
---input="1x64xi64"
---input="1x64xi64"
---input="1x64xi64"
---parameters=splats.irpa
