--- iree_tests/pytorch/models/sdxl-prompt-encoder-tank/splats.irpa
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:018b337cbbc5f66d0791b3487f80bf8f5f1c3634f567f89bc160c66477a462ca
-size 118784
