--- iree_tests/pytorch/models/sdxl-scheduled-unet-3-tank/model.mlirbc
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:81e934a304c92495182703ba8f1c3e31d9f5ce2c9a4d1f1e7dfbd63c2e3a804f
-size 1004980
