--- iree_tests/pytorch/models/sdxl-scheduled-unet-3-tank/real_weights_data_flags.txt
@@ -1,1 +1,0 @@
-
