--- iree_tests/pytorch/models/sdxl-scheduled-unet-3-tank/splat_data_flags.txt
@@ -1,5 +1,0 @@
---input="1x4x128x128xf16"
---input="2x64x2048xf16"
---input="2x1280xf16"
---input="1xf16"
---parameters=splats.irpa
