--- iree_tests/pytorch/models/sdxl-scheduled-unet-3-tank/splats.irpa
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d3dd8f1ee797944c3ab3286858b89257dbbb55bb7b894c82437d1a0d94b9f1a9
-size 270336
