--- iree_tests/pytorch/models/sdxl-vae-decode-tank/model.mlirbc
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5598fc33d35b2eb32e88b5136cfe220146f4b6ef749e26aba56435a66aa34b97
-size 100423
