--- iree_tests/pytorch/models/sdxl-vae-decode-tank/splat_data_flags.txt
@@ -1,2 +1,0 @@
---input="1x4x128x128xf16"
---parameters=splats.irpa
