--- iree_tests/pytorch/models/sdxl-vae-decode-tank/splats.irpa
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:01bb22697cfde0ac4ffdb1440b92ab98d6f757dc92b5399febb868fbef6f7641
-size 36864
