--- iree_tests/pytorch/models/sdxl-vae-decode-tank/test_cases.json
@@ -1,19 +1,0 @@
-{
-  "file_format": "test_cases_v0",
-  "test_cases": [
-    {
-      "name": "splats",
-      "runtime_flagfile": "splat_data_flags.txt",
-      "remote_files": []
-    },
-    {
-      "name": "real_weights",
-      "runtime_flagfile": "real_weights_data_flags.txt",
-      "remote_files": [
-        "https://sharkpublic.blob.core.windows.net/sharkpublic/sai/sdxl-vae-decode/inference_input.0.bin",
-        "https://sharkpublic.blob.core.windows.net/sharkpublic/sai/sdxl-vae-decode/inference_output.0.bin",
-        "https://sharkpublic.blob.core.windows.net/sharkpublic/sai/sdxl-vae-decode/real_weights.irpa"
-      ]
-    }
-  ]
-}
