+++ iree_tests/requirements.txt
@@ -6,6 +6,7 @@ pytest-timeout
 pytest-xdist
 pyjson5
 pytest-check
+pytest-dependency
 
 # Remote file downloading
 azure-storage-blob
